==> code/01b_set_flextable_defaults.R <==
# ---------------------------------------------------------------------------- #
# Set "flextable" Package Defaults
# Author: Jeremy W. Eberle
# ---------------------------------------------------------------------------- #

# Set defaults for "flextable" package (mimic those of MS Word) that can be
# sourced in other scripts

set_flextable_defaults(font.size = 12, font.family = "Times New Roman", font.color = "black", 
                       border.width = 0.5, border.color = "black",
                       padding.bottom = 0, padding.top = 0, 
                       padding.left = 0.08, padding.right = 0.08,
                       line_spacing = 1)
